+++ packages/pages/src/generate/templates/templates.ts
@@ -1,21 +1,19 @@
 import { ProjectStructure } from "../../common/src/project/structure.js";
 import { getTemplateFilepaths } from "../../common/src/template/internal/getTemplateFilepaths.js";
-import { loadTemplateModules } from "../../common/src/template/internal/loader.js";
 import { createTemplatesJson } from "./createTemplatesJson.js";
 import { Command } from "commander";
 
-const handler = async ({ scope }: { scope: string }): Promise<void> => {
+export const templatesHandler = async ({
+  scope,
+}: {
+  scope: string;
+}): Promise<void> => {
   const projectStructure = await ProjectStructure.init({ scope });
   const templateFilepaths = getTemplateFilepaths(
     projectStructure.getTemplatePaths()
   );
-  const templateModules = await loadTemplateModules(
-    templateFilepaths,
-    true,
-    false
-  );
 
-  createTemplatesJson(templateModules, projectStructure);
+  await createTemplatesJson(templateFilepaths, projectStructure, "TEMPLATES");
 };
 
 export const templatesCommand = (program: Command) => {
@@ -26,5 +24,5 @@ export const templatesCommand = (program: Command) => {
       "--scope <string>",
       "The subfolder to scope the served templates from"
     )
-    .action(handler);
+    .action(templatesHandler);
 };