+++ packages/pages/src/util/runSubprocess.test.ts
@@ -7,7 +7,12 @@ describe("runSubprocess", () => {
   });
 
   it("returns non-zero when subprocess fails", async () => {
-    const exitCode = await runSubprocess("echo", ["$((0/0))"]);
-    expect(exitCode).not.toEqual(0);
+    if (process.platform === "win32") {
+      const exitCode = await runSubprocess("powershell.exe", ["0/0"]);
+      expect(exitCode).not.toEqual(0);
+    } else {
+      const exitCode = await runSubprocess("echo", ["$((0/0))"]);
+      expect(exitCode).not.toEqual(0);
+    }
   });
 });